+++ src/components/Agent/index.js
@@ -2,8 +2,8 @@
 import React, {Component} from 'react';
 import { compose } from 'recompose';
 
-import { withFirebase } from '../Firebase';
-import { withAuthorization, withEmailVerification } from '../Session';
+import { withFirebase } from '../../server/Firebase/index';
+import { withAuthorization, withEmailVerification } from '../../server/Session/index';
 
 import * as ROLES from '../../constants/roles';
 const AgentPage = () => (
