+++ src/components/PasswordChange/index.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
 import {Link, withRouter } from 'react-router-dom';
 
-import {withFirebase} from '../Firebase';
+import {withFirebase} from '../../server/Firebase/index';
 import * as ROUTES from '../../constants/routes';
 
 const PasswordChangePage = () => (
