+++ src/index.js
@@ -5,7 +5,7 @@ import './index.css';
 import * as serviceWorker from './serviceWorker';
 
 import App from './components/App';
-import Firebase, {FirebaseContext} from './components/Firebase';
+import Firebase, {FirebaseContext} from './server/Firebase/index';
 
 import 'bootstrap/dist/css/bootstrap.css';
