+++ src/server/Session/withAuthorization.js
@@ -3,7 +3,7 @@ import { withRouter } from 'react-router-dom';
 import {compose} from 'recompose';
 
 import AuthUserContext from './context';
-import {withFirebase} from '../Firebase';
+import {withFirebase} from '../Firebase/index';
 import * as ROUTES from '../../constants/routes';
 /*
 Authorization check and rediriect unauthorized user
